refactor: 💡 move theme folder out of ui folder
