+++ app/src/main/java/com/rickyhu/hushkeyboard/keyboard/ui/buttons/ControlKeyButton.kt
@@ -9,9 +9,9 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.rickyhu.hushkeyboard.ui.theme.DarkSecondary
-import com.rickyhu.hushkeyboard.ui.theme.HushKeyboardTheme
-import com.rickyhu.hushkeyboard.ui.theme.LightSecondary
+import com.rickyhu.hushkeyboard.theme.DarkSecondary
+import com.rickyhu.hushkeyboard.theme.HushKeyboardTheme
+import com.rickyhu.hushkeyboard.theme.LightSecondary
 
 @Composable
 fun ControlKeyButton(
