+++ app/src/main/java/com/rickyhu/hushkeyboard/main/MainApp.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.rickyhu.hushkeyboard.data.ThemeOption
 import com.rickyhu.hushkeyboard.navigation.AppNavHost
-import com.rickyhu.hushkeyboard.ui.theme.HushKeyboardTheme
+import com.rickyhu.hushkeyboard.theme.HushKeyboardTheme
 
 @Composable
 fun MainApp(
