+++ app/src/main/java/com/rickyhu/hushkeyboard/settings/ui/WideNotationOptionDropdownItem.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.rickyhu.hushkeyboard.R
 import com.rickyhu.hushkeyboard.data.WideNotationOption
-import com.rickyhu.hushkeyboard.ui.theme.HushKeyboardTheme
+import com.rickyhu.hushkeyboard.theme.HushKeyboardTheme
 
 @Composable
 fun WideNotationOptionDropdownItem(
