+++ app/src/main/java/com/rickyhu/hushkeyboard/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.rickyhu.hushkeyboard.ui.theme
+package com.rickyhu.hushkeyboard.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
